chore: Bump version to 0.1.6 and setup release-please
- Bumped the version from 0.1.5 to 0.1.6 in `README.md`, `pyproject.toml`, and all `Cargo.toml` files.
- Added `.github/workflows/release-please.yml` to automatically manage semantic versioning and releases via pull requests.
- Added `.release-please-config.json` and `.release-please-manifest.json` targeting the core package, python bindings, wasm bindings, and `README.md` for coordinated version bumps.

+++ crates/fuzzgpu-core/src/gpu.rs
@@ -601,7 +601,7 @@ impl GpuEngine {
             device,
             queue,
             info,
-            max_buffer_size: target_buffer_size as u64,
+            max_buffer_size: target_buffer_size,
             // wgpu 30 widened this limit to u64; the 128 MiB cap fits u32.
             max_storage_buffer_binding_size: target_storage_size as u32,
             dispatch_lock: std::sync::Mutex::new(()),
